[Refactor] : Requestable, NetworkError 파일 생성, 코드 분리
Related to: #6
